Adds blocked users view
also refactors the setting view
