+++ bookwyrm/templates/change_password.html
@@ -1,0 +1,19 @@
+{% extends 'preferences_layout.html' %}
+{% block header %}
+Change Password
+{% endblock %}
+
+{% block panel %}
+<form name="edit-profile" action="/change-password/" method="post" enctype="multipart/form-data">
+    {% csrf_token %}
+    <div class="block">
+        <label class="label" for="id_password">New password:</label>
+        <input type="password" name="password" maxlength="128" class="input" required="" id="id_password">
+    </div>
+    <div class="block">
+        <label class="label" for="id_confirm_password">Confirm password:</label>
+        <input type="password" name="confirm-password" maxlength="128" class="input" required="" id="id_confirm_password">
+    </div>
+    <button class="button is-primary" type="submit">Change password</button>
+</form>
+{% endblock %}
