+++ bookwyrm/templates/snippets/block_button.html
@@ -1,11 +1,11 @@
 {% if not user in request.user.blocks.all %}
 <form name="blocks" method="post" action="/block/{{ user.id }}">
     {% csrf_token %}
-    <button class="button is-danger is-light is-small is-fullwidth" type="submit">Block</button>
+    <button class="button is-danger is-light is-small {{ class }}" type="submit">Block</button>
 </form>
 {% else %}
 <form name="unblocks" method="post" action="/unblock/{{ user.id }}">
     {% csrf_token %}
-    <button class="button is-small is-fullwidth" type="submit">Un-block</button>
+    <button class="button is-small {{ class }}" type="submit">Un-block</button>
 </form>
 {% endif %}
