+++ bookwyrm/tests/views/test_authentication.py
@@ -42,83 +42,6 @@ class AuthenticationViews(TestCase):
         self.assertEqual(result.status_code, 302)
 
 
-    def test_password_reset_request(self):
-        ''' there are so many views, this just makes sure it LOADS '''
-        view = views.PasswordResetRequest.as_view()
-        request = self.factory.get('')
-        request.user = self.local_user
-
-        result = view(request)
-        self.assertIsInstance(result, TemplateResponse)
-        self.assertEqual(result.template_name, 'password_reset_request.html')
-        self.assertEqual(result.status_code, 200)
-
-
-    def test_password_reset_request_post(self):
-        ''' send 'em an email '''
-        request = self.factory.post('', {'email': 'aa@bb.ccc'})
-        view = views.PasswordResetRequest.as_view()
-        resp = view(request)
-        self.assertEqual(resp.status_code, 302)
-
-        request = self.factory.post('', {'email': 'mouse@mouse.com'})
-        with patch('bookwyrm.emailing.send_email.delay'):
-            resp = view(request)
-        self.assertEqual(resp.template_name, 'password_reset_request.html')
-
-        self.assertEqual(
-            models.PasswordReset.objects.get().user, self.local_user)
-
-    def test_password_reset(self):
-        ''' there are so many views, this just makes sure it LOADS '''
-        view = views.PasswordReset.as_view()
-        code = models.PasswordReset.objects.create(user=self.local_user)
-        request = self.factory.get('')
-        request.user = self.anonymous_user
-        result = view(request, code.code)
-        self.assertIsInstance(result, TemplateResponse)
-        self.assertEqual(result.template_name, 'password_reset.html')
-        self.assertEqual(result.status_code, 200)
-
-
-    def test_password_reset_post(self):
-        ''' reset from code '''
-        view = views.PasswordReset.as_view()
-        code = models.PasswordReset.objects.create(user=self.local_user)
-        request = self.factory.post('', {
-            'password': 'hi',
-            'confirm-password': 'hi'
-        })
-        with patch('bookwyrm.views.password.login'):
-            resp = view(request, code.code)
-        self.assertEqual(resp.status_code, 302)
-        self.assertFalse(models.PasswordReset.objects.exists())
-
-    def test_password_reset_wrong_code(self):
-        ''' reset from code '''
-        view = views.PasswordReset.as_view()
-        models.PasswordReset.objects.create(user=self.local_user)
-        request = self.factory.post('', {
-            'password': 'hi',
-            'confirm-password': 'hi'
-        })
-        resp = view(request, 'jhgdkfjgdf')
-        self.assertEqual(resp.template_name, 'password_reset.html')
-        self.assertTrue(models.PasswordReset.objects.exists())
-
-    def test_password_reset_mismatch(self):
-        ''' reset from code '''
-        view = views.PasswordReset.as_view()
-        code = models.PasswordReset.objects.create(user=self.local_user)
-        request = self.factory.post('', {
-            'password': 'hi',
-            'confirm-password': 'hihi'
-        })
-        resp = view(request, code.code)
-        self.assertEqual(resp.template_name, 'password_reset.html')
-        self.assertTrue(models.PasswordReset.objects.exists())
-
-
     def test_register(self):
         ''' create a user '''
         view = views.Register.as_view()
@@ -274,29 +197,3 @@ class AuthenticationViews(TestCase):
         with self.assertRaises(Http404):
             response = view(request)
         self.assertEqual(models.User.objects.count(), 2)
-
-
-    def test_password_change(self):
-        ''' change password '''
-        view = views.ChangePassword.as_view()
-        password_hash = self.local_user.password
-        request = self.factory.post('', {
-            'password': 'hi',
-            'confirm-password': 'hi'
-        })
-        request.user = self.local_user
-        with patch('bookwyrm.views.password.login'):
-            view(request)
-        self.assertNotEqual(self.local_user.password, password_hash)
-
-    def test_password_change_mismatch(self):
-        ''' change password '''
-        view = views.ChangePassword.as_view()
-        password_hash = self.local_user.password
-        request = self.factory.post('', {
-            'password': 'hi',
-            'confirm-password': 'hihi'
-        })
-        request.user = self.local_user
-        view(request)
-        self.assertEqual(self.local_user.password, password_hash)
